# openapi/tokenpilot.openapi.yaml
openapi: 3.1.0
info:
  title: TokenPilot Local Control Plane API
  version: 0.1.0-alpha
  description: >
    Local-first automation surface for TokenPilot. Phase 2 supports ChatGPT direct-drive
    operations for small trusted edits plus asynchronous Codex jobs for complex work. The
    full HTTPS / Custom GPT Actions automation loop remains under validation.
servers:
  - url: https://tokenpilot.example.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ContinuityReadResponse:
      type: object
      additionalProperties: true
      properties:
        ok:
          type: boolean
          const: true
      required: [ok]
    ContinuityMutationResponse:
      type: object
      additionalProperties: true
      properties:
        ok:
          type: boolean
          const: true
        replayed:
          type: boolean
      required: [ok, replayed]
    ContinuityTaskCreatePayload:
      type: object
      additionalProperties: false
      properties:
        projectId:
          type: string
        workspaceId:
          type: string
        specId:
          oneOf:
            - type: string
            - type: "null"
        planId:
          oneOf:
            - type: string
            - type: "null"
        parentTaskId:
          oneOf:
            - type: string
            - type: "null"
        title:
          type: string
          minLength: 1
          maxLength: 240
        goal:
          type: string
          minLength: 1
          maxLength: 12000
        priority:
          type: string
          enum: [low, normal, high, critical]
          default: normal
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [projectId, workspaceId, title, goal, idempotencyKey]
    ContinuityTaskSubmitReviewPayload:
      type: object
      additionalProperties: false
      properties:
        taskId:
          type: string
        expectedRevision:
          type: integer
          minimum: 1
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [taskId, expectedRevision, idempotencyKey]
    ContinuityTaskSubmitReviewResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        task:
          $ref: '#/components/schemas/ContinuityTaskRecord'
        evidenceBundle:
          $ref: '#/components/schemas/ContinuityEvidenceBundleRecord'
        replayed:
          type: boolean
      required: [ok, task, evidenceBundle, replayed]
    ContinuityTaskCompletePayload:
      type: object
      additionalProperties: false
      properties:
        taskId:
          type: string
        expectedRevision:
          type: integer
          minimum: 1
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [taskId, expectedRevision, idempotencyKey]
    ContinuityTaskCompleteResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        task:
          $ref: '#/components/schemas/ContinuityTaskRecord'
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeDevelopmentSessionRecord'
        handoff:
          $ref: '#/components/schemas/ContinuityHandoffRecord'
        evidenceBundle:
          $ref: '#/components/schemas/ContinuityEvidenceBundleRecord'
        replayed:
          type: boolean
      required: [ok, task, sessions, handoff, evidenceBundle, replayed]
    ContinuitySessionStartPayload:
      type: object
      additionalProperties: false
      properties:
        taskId:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 240
        mode:
          type: string
          enum: [chat-direct, codex-session, async-agent]
        expectedTaskRevision:
          type: integer
          minimum: 1
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [taskId, title, mode, expectedTaskRevision, idempotencyKey]
    ContinuityLeaseAcquirePayload:
      type: object
      additionalProperties: false
      properties:
        sessionId:
          type: string
        holderId:
          type: string
        expiresAt:
          type: string
          format: date-time
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [sessionId, holderId, expiresAt, idempotencyKey]
    ContinuityLeaseReleasePayload:
      type: object
      additionalProperties: false
      properties:
        leaseId:
          type: string
        sessionId:
          type: string
        holderId:
          type: string
        expectedRevision:
          type: integer
          minimum: 1
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [leaseId, sessionId, holderId, expectedRevision, idempotencyKey]
    ContinuityHandoffPreparePayload:
      type: object
      additionalProperties: false
      properties:
        taskId:
          type: string
        sessionId:
          type: string
        toMode:
          type: string
          enum: [chat-direct, codex-session, async-agent, unassigned]
          default: unassigned
        goal:
          type: string
          minLength: 1
          maxLength: 12000
        completedItems:
          type: array
          maxItems: 200
          items:
            type: string
        pendingItems:
          type: array
          maxItems: 200
          items:
            type: string
        changedFiles:
          type: array
          maxItems: 500
          items:
            type: string
        risks:
          type: array
          maxItems: 200
          items:
            type: string
        nextAction:
          type: string
          minLength: 1
          maxLength: 4000
        gitHead:
          oneOf:
            - type: string
            - type: "null"
        gitBranch:
          oneOf:
            - type: string
            - type: "null"
        gitDirty:
          type: boolean
        diffArtifactId:
          oneOf:
            - type: string
            - type: "null"
        evidenceBundleId:
          oneOf:
            - type: string
            - type: "null"
        expectedTaskRevision:
          type: integer
          minimum: 1
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [taskId, sessionId, goal, nextAction, gitDirty, expectedTaskRevision, idempotencyKey]
    ContinuityHandoffAcceptPayload:
      type: object
      additionalProperties: false
      properties:
        handoffId:
          type: string
        expectedRevision:
          type: integer
          minimum: 1
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [handoffId, expectedRevision, idempotencyKey]
    ContinuityHandoffCancelPayload:
      $ref: '#/components/schemas/ContinuityHandoffAcceptPayload'
    ContinuityHandoffForkPayload:
      type: object
      additionalProperties: false
      properties:
        handoffId:
          type: string
        expectedRevision:
          type: integer
          minimum: 1
        title:
          type: string
          minLength: 1
          maxLength: 240
        sessionTitle:
          type: string
          minLength: 1
          maxLength: 240
        mode:
          type: string
          enum: [chat-direct, codex-session, async-agent]
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [handoffId, expectedRevision, title, sessionTitle, idempotencyKey]
    ContinuityEvidenceRecordPayload:
      type: object
      additionalProperties: false
      properties:
        taskId:
          type: string
        sessionId:
          type: string
        bundleId:
          type: string
        kind:
          type: string
          enum: [command, test, build, lint, typecheck, diff, review, screenshot, manual]
        label:
          type: string
          minLength: 1
          maxLength: 240
        status:
          type: string
          enum: [passed, failed, skipped, not-run]
        required:
          type: boolean
          default: false
        command:
          oneOf:
            - type: string
            - type: "null"
        exitCode:
          oneOf:
            - type: integer
            - type: "null"
        artifactId:
          oneOf:
            - type: string
            - type: "null"
        summary:
          type: string
          maxLength: 8000
        startedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        completedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        expectedTaskRevision:
          type: integer
          minimum: 1
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [taskId, sessionId, kind, label, status, idempotencyKey]
    ContinuityProjectRecord:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        slug:
          type: string
        displayName:
          type: string
        defaultWorkspaceId:
          oneOf:
            - type: string
            - type: "null"
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        revision:
          type: integer
          minimum: 1
      required: [id, slug, displayName, defaultWorkspaceId, status, createdAt, updatedAt, revision]
...
    ContinuityTaskRecord:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        projectId:
          type: string
        workspaceId:
          type: string
        specId:
          oneOf:
            - type: string
            - type: "null"
        planId:
          oneOf:
            - type: string
            - type: "null"
        parentTaskId:
          oneOf:
            - type: string
            - type: "null"
        title:
          type: string
        goal:
          type: string
        status:
          type: string
          enum: [backlog, ready, in-progress, blocked, review, completed, cancelled]
        priority:
          type: string
          enum: [low, normal, high, critical]
        activeSessionId:
          oneOf:
            - type: string
            - type: "null"
        latestHandoffId:
          oneOf:
            - type: string
            - type: "null"
        latestEvidenceBundleId:
          oneOf:
            - type: string
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        revision:
          type: integer
          minimum: 1
      required: [id, projectId, workspaceId, specId, planId, parentTaskId, title, goal, status, priority, activeSessionId, latestHandoffId, latestEvidenceBundleId, createdAt, updatedAt, revision]
...
    ContinuityEvidenceBundleRecord:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        taskId:
          type: string
        sessionId:
          type: string
        status:
          type: string
          enum: [collecting, complete, incomplete]
        requiredItemCount:
          type: integer
          minimum: 0
        passedItemCount:
          type: integer
          minimum: 0
        failedItemCount:
          type: integer
          minimum: 0
        skippedItemCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        completedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        revision:
          type: integer
          minimum: 1
      required: [id, taskId, sessionId, status, requiredItemCount, passedItemCount, failedItemCount, skippedItemCount, createdAt, completedAt, revision]
...
    WorkspaceContinuitySnapshot:
      type: object
      additionalProperties: false
      properties:
        project:
          $ref: '#/components/schemas/ContinuityProjectRecord'
        workspace:
          $ref: '#/components/schemas/ContinuityWorkspaceRecord'
        activeLease:
          oneOf:
            - $ref: '#/components/schemas/ContinuityWriterLeaseRecord'
            - type: "null"
        readOnly:
          type: boolean
        readOnlyReason:
          oneOf:
            - type: string
              enum: [active-writer]
            - type: "null"
        git:
          $ref: '#/components/schemas/WorkspaceGitProjection'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceTaskContinuityProjection'
        pendingApprovals:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeApprovalRecord'
      required: [project, workspace, activeLease, readOnly, readOnlyReason, git, tasks, pendingApprovals]
    WorkspaceContinuitySnapshotResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        snapshot:
          $ref: '#/components/schemas/WorkspaceContinuitySnapshot'
      required: [ok, snapshot]
...
    RuntimeDevelopmentSessionRecord:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        projectId:
          type: string
        workspaceId:
          type: string
        taskId:
          type: string
        title:
          type: string
        mode:
          type: string
          enum: [chat-direct, codex-session, async-agent]
        status:
          type: string
          enum: [idle, running, waiting-approval, handoff-ready, completed, failed]
        activeRuntimeBindingId:
          oneOf:
            - type: string
            - type: "null"
        startedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        endedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        revision:
          type: integer
          minimum: 1
      required:
        - id
        - projectId
        - workspaceId
        - taskId
        - title
        - mode
        - status
        - activeRuntimeBindingId
        - startedAt
        - updatedAt
        - endedAt
        - revision
...
paths:
...
  /api/continuity/projects:
    get:
      operationId: listContinuityProjects
      summary: List continuity projects
      description: Synchronizes allowlisted repo mappings into durable TokenPilot project and workspace identities. Local absolute paths are omitted from the response.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, archived]
      responses:
        "200":
          description: Public-safe project projections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityReadResponse'
  /api/continuity/projects/{projectId}:
    get:
      operationId: getContinuityProject
      summary: Read continuity project
      description: Returns one durable project and its public-safe workspace projections.
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityReadResponse'
  /api/continuity/workspaces/{workspaceId}/snapshot:
    get:
      operationId: getWorkspaceContinuitySnapshot
      summary: Read workspace continuity snapshot
      description: Returns public-safe Git, active writer, task, session, handoff, evidence, and pending approval state for one workspace.
      security:
        - bearerAuth: []
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workspace continuity snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceContinuitySnapshotResponse'
  /api/continuity/tasks:
    post:
      operationId: createContinuityTask
      summary: Create continuity task
      description: Creates a durable project task. The idempotency key prevents duplicate task creation across REST and MCP retries.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityTaskCreatePayload'
      responses:
        "200":
          description: Created or replayed task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityMutationResponse'
  /api/continuity/tasks/submit-review:
    post:
      operationId: submitContinuityTaskReview
      summary: Submit continuity task for review
      description: Finalizes passed required evidence and moves an in-progress or blocked task into review through one idempotent transition.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityTaskSubmitReviewPayload'
      responses:
        "200":
          description: Submitted or replayed task review transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityTaskSubmitReviewResponse'
        "409":
          description: Review evidence requirements are not satisfied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/continuity/tasks/complete:
    post:
      operationId: completeContinuityTask
      summary: Complete continuity task
      description: Completes a review task only after accepted handoff, required evidence, writer, runtime run, and approval checks pass.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityTaskCompletePayload'
      responses:
        "200":
          description: Completed or replayed task transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityTaskCompleteResponse'
        "409":
          description: Completion requirements are not satisfied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/continuity/tasks/{taskId}:
    get:
      operationId: getContinuityTask
      summary: Read continuity task
      description: Returns one durable TokenPilot task by id.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityReadResponse'
  /api/continuity/sessions/start:
    post:
      operationId: startContinuitySession
      summary: Start development session
      description: Starts a Chat Direct, Codex Session, or Async Agent session and binds it to the expected task revision.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuitySessionStartPayload'
      responses:
        "200":
          description: Started or replayed session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityMutationResponse'
  /api/continuity/sessions/{sessionId}:
    get:
      operationId: getContinuitySession
      summary: Read development session
      description: Returns one durable TokenPilot development session by id.
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityReadResponse'
  /api/continuity/leases/acquire:
    post:
      operationId: acquireContinuityLease
      summary: Acquire workspace writer lease
      description: Acquires the only active writer lease allowed for a workspace. Another active session receives a conflict.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityLeaseAcquirePayload'
      responses:
        "200":
          description: Acquired or replayed writer lease
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityMutationResponse'
  /api/continuity/leases/release:
    post:
      operationId: releaseContinuityLease
      summary: Release workspace writer lease
      description: Releases an active writer lease using lease identity, holder identity, revision, and idempotency control.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityLeaseReleasePayload'
      responses:
        "200":
          description: Released or replayed writer lease
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityMutationResponse'
  /api/continuity/handoffs/prepare:
    post:
      operationId: prepareContinuityHandoff
      summary: Prepare development handoff
      description: Creates a ready handoff checkpoint with goal, progress, pending work, changed files, risks, Git state, evidence, and next action.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityHandoffPreparePayload'
      responses:
        "200":
          description: Ready or replayed handoff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityMutationResponse'
  /api/continuity/handoffs/cancel:
    post:
      operationId: cancelContinuityHandoff
      summary: Cancel development handoff
      description: Supersedes one ready handoff with optimistic revision and idempotency controls.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityHandoffCancelPayload'
      responses:
        "200":
          description: Cancelled or replayed handoff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityMutationResponse'
  /api/continuity/handoffs/fork:
    post:
      operationId: forkContinuityHandoff
      summary: Fork development handoff
      description: Consumes one ready handoff by creating a child task and target-mode development session in one transaction.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityHandoffForkPayload'
      responses:
        "200":
          description: Forked or replayed handoff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityMutationResponse'
  /api/continuity/handoffs/accept:
    post:
      operationId: acceptContinuityHandoff
      summary: Accept development handoff
      description: Accepts one ready handoff with optimistic revision and idempotency controls.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityHandoffAcceptPayload'
      responses:
        "200":
          description: Accepted or replayed handoff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityMutationResponse'
  /api/continuity/evidence/record:
    post:
      operationId: recordContinuityEvidence
      summary: Record verification evidence
      description: Records structured verification evidence for a task session and optionally creates a new evidence bundle.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinuityEvidenceRecordPayload'
      responses:
        "200":
          description: Recorded or replayed evidence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContinuityMutationResponse'
...